# Let all Init_xxx and extension APIs frequently called from init code paths be considered cold