--- a/array.c
+++ b/array.c
@@ -6250,6 +6250,7 @@ rb_ary_sum(int argc, VALUE *argv, VALUE ary)
  *
  */
 
+COLDFUNC(void Init_Array(void));
 void
 Init_Array(void)
 {