--- a/complex.c
+++ b/complex.c
@@ -2137,6 +2137,7 @@ float_arg(VALUE self)
  *    Complex(1, 1) / 2    #=> ((1/2)+(1/2)*i)
  *    Complex(1, 1) / 2.0  #=> (0.5+0.5i)
  */
+COLDFUNC(void Init_Complex(void));
 void
 Init_Complex(void)
 {