--- a/enumerator.c
+++ b/enumerator.c
@@ -2778,6 +2778,7 @@ arith_seq_size(VALUE self)
     return len;
 }
 
+COLDFUNC(void InitVM_Enumerator(void));
 void
 InitVM_Enumerator(void)
 {
@@ -2875,6 +2876,7 @@ InitVM_Enumerator(void)
 }
 
 #undef rb_intern
+COLDFUNC(void Init_Enumerator(void));
 void
 Init_Enumerator(void)
 {