--- a/ext/digest/digest.c
+++ b/ext/digest/digest.c
@@ -23,7 +23,7 @@ static VALUE rb_cDigest_Base;
 static ID id_reset, id_update, id_finish, id_digest, id_hexdigest, id_digest_length;
 static ID id_metadata;
 
-RUBY_EXTERN void Init_digest_base(void);
+COLDFUNC(RUBY_EXTERN void Init_digest_base(void));
 
 /*
  * Document-module: Digest
@@ -725,6 +725,7 @@ rb_digest_base_block_length(VALUE self)
     return INT2NUM(algo->block_len);
 }
 
+COLDFUNC(void Init_digest(void));
 void
 Init_digest(void)
 {