--- a/ext/digest/sha2/sha2init.c
+++ b/ext/digest/sha2/sha2init.c
@@ -31,6 +31,7 @@ FOREACH_BITLEN(DEFINE_ALGO_METADATA)
  * Secure Hash Algorithm(s) by NIST (the US' National Institute of
  * Standards and Technology), described in FIPS PUB 180-2.
  */
+COLDFUNC(void Init_sha2(void));
 void
 Init_sha2(void)
 {