--- a/ext/openssl/ossl_digest.h
+++ b/ext/openssl/ossl_digest.h
@@ -15,6 +15,6 @@ extern VALUE eDigestError;
 
 const EVP_MD *ossl_evp_get_digestbyname(VALUE);
 VALUE ossl_digest_new(const EVP_MD *);
-void Init_ossl_digest(void);
+COLDFUNC(void Init_ossl_digest(void));
 
 #endif /* _OSSL_DIGEST_H_ */