--- a/ext/openssl/ossl_hmac.h
+++ b/ext/openssl/ossl_hmac.h
@@ -13,6 +13,6 @@
 extern VALUE cHMAC;
 extern VALUE eHMACError;
 
-void Init_ossl_hmac(void);
+COLDFUNC(void Init_ossl_hmac(void));
 
 #endif /* _OSSL_HMAC_H_ */