--- a/ext/openssl/ossl_kdf.h
+++ b/ext/openssl/ossl_kdf.h
@@ -1,6 +1,6 @@
 #if !defined(OSSL_KDF_H)
 #define OSSL_KDF_H
 
-void Init_ossl_kdf(void);
+COLDFUNC(void Init_ossl_kdf(void));
 
 #endif