--- a/ext/openssl/ossl_pkcs12.h
+++ b/ext/openssl/ossl_pkcs12.h
@@ -8,6 +8,6 @@
 extern VALUE cPKCS12;
 extern VALUE ePKCS12Error;
 
-void Init_ossl_pkcs12(void);
+COLDFUNC(void Init_ossl_pkcs12(void));
 
 #endif /* _OSSL_PKCS12_H_ */