--- a/ext/psych/psych_emitter.h
+++ b/ext/psych/psych_emitter.h
@@ -3,6 +3,6 @@
 
 #include <psych.h>
 
-void Init_psych_emitter(void);
+COLDFUNC(void Init_psych_emitter(void));
 
 #endif