--- a/ext/psych/psych_parser.h
+++ b/ext/psych/psych_parser.h
@@ -1,6 +1,6 @@
 #ifndef PSYCH_PARSER_H
 #define PSYCH_PARSER_H
 
-void Init_psych_parser(void);
+COLDFUNC(void Init_psych_parser(void));
 
 #endif