--- a/ext/psych/psych_to_ruby.h
+++ b/ext/psych/psych_to_ruby.h
@@ -3,6 +3,6 @@
 
 #include <psych.h>
 
-void Init_psych_to_ruby(void);
+COLDFUNC(void Init_psych_to_ruby(void));
 
 #endif