--- a/ext/psych/psych_yaml_tree.h
+++ b/ext/psych/psych_yaml_tree.h
@@ -3,6 +3,6 @@
 
 #include <psych.h>
 
-void Init_psych_yaml_tree(void);
+COLDFUNC(void Init_psych_yaml_tree(void));
 
 #endif