--- a/ext/win32ole/win32ole.c
+++ b/ext/win32ole/win32ole.c
@@ -3975,6 +3975,7 @@ check_nano_server(void)
 }
 
 
+COLDFUNC(void Init_win32ole(void));
 void
 Init_win32ole(void)
 {