--- a/ext/win32ole/win32ole_event.h
+++ b/ext/win32ole/win32ole_event.h
@@ -1,6 +1,6 @@
 #ifndef WIN32OLE_EVENT_H
 #define WIN32OLE_EVENT_H 1
 
-void Init_win32ole_event(void);
+COLDFUNC(void Init_win32ole_event(void));
 
 #endif