--- a/ext/win32ole/win32ole_method.h
+++ b/ext/win32ole/win32ole_method.h
@@ -12,5 +12,5 @@ VALUE folemethod_s_allocate(VALUE klass);
 VALUE ole_methods_from_typeinfo(ITypeInfo *pTypeInfo, int mask);
 VALUE create_win32ole_method(ITypeInfo *pTypeInfo, VALUE name);
 struct olemethoddata *olemethod_data_get_struct(VALUE obj);
-void Init_win32ole_method(void);
+COLDFUNC(void Init_win32ole_method(void));
 #endif