--- a/ext/win32ole/win32ole_param.h
+++ b/ext/win32ole/win32ole_param.h
@@ -2,7 +2,7 @@
 #define WIN32OLE_PARAM_H
 
 VALUE create_win32ole_param(ITypeInfo *pTypeInfo, UINT method_index, UINT index, VALUE name);
-void Init_win32ole_param(void);
+COLDFUNC(void Init_win32ole_param(void));
 
 #endif