--- a/ext/win32ole/win32ole_variant.h
+++ b/ext/win32ole/win32ole_variant.h
@@ -3,7 +3,7 @@
 
 VALUE cWIN32OLE_VARIANT;
 void ole_variant2variant(VALUE val, VARIANT *var);
-void Init_win32ole_variant(void);
+COLDFUNC(void Init_win32ole_variant(void));
 
 #endif