--- a/ext/win32ole/win32ole_variant_m.h
+++ b/ext/win32ole/win32ole_variant_m.h
@@ -2,6 +2,7 @@
 #define WIN32OLE_VARIANT_M_H 1
 
 VALUE mWIN32OLE_VARIANT;
+COLDFUNC(void Init_win32ole_variant_m(void));
 void Init_win32ole_variant_m(void);
 
 #endif