… internal.h
… load.c
… marshal.c
… math.c
… numeric.c
… object.c
… pack.c
… parse.y
… proc.c
… process.c
… random.c
… range.c
… rational.c
… re.c
… ruby.c
… safe.c
… signal.c
… string.c
… struct.c
… symbol.c
… template/extinit.c.tmpl
… template/id.c.tmpl
… template/limits.c.tmpl
… template/prelude.c.tmpl
… template/sizes.c.tmpl
… thread.c
… thread_sync.c
… thread_win32.c
… time.c
… transcode.c