@methodmissing 2018-08-18
For benchmarks, I'm getting a lot of noise for optcarrot and ruby benchmarks and consider this specific change as inconclusive with regards to changing anything performance wise:

```
lourens@CarbonX1:~/src/optcarrot$ benchmark-driver -e "cold::~/src/ruby/ruby/ruby -I~/src/ruby/ruby/lib -I~/src/ruby/ruby/. -I~/src/ruby/ruby/.ext/x86_64-linux" -e "trunk::~/src/ruby/trunk/ruby -I~/src/ruby/trunk/lib -I~/src/ruby/trunk/. -I~/src/ruby/trunk/.ext/x86_64-linux" --repeat-count 1 --repeat-result average -vv benchmark.yml
cold: ruby 2.6.0dev (2018-08-19 coldfunc-init 64464) [x86_64-linux]
last_commit=Let all Init_xxx and extension APIs frequently called from init code paths be considered cold
trunk: ruby 2.6.0dev (2018-08-18 trunk 64448) [x86_64-linux]
Calculating -------------------------------------
                           cold       trunk 
           optcarrot     45.113      43.008 fps

Comparison:
                        optcarrot
                cold:        45.1 fps 
               trunk:        43.0 fps - 1.05x  slower

lourens@CarbonX1:~/src/optcarrot$ benchmark-driver -e "cold::~/src/ruby/ruby/ruby -I~/src/ruby/ruby/lib -I~/src/ruby/ruby/. -I~/src/ruby/ruby/.ext/x86_64-linux" -e "trunk::~/src/ruby/trunk/ruby -I~/src/ruby/trunk/lib -I~/src/ruby/trunk/. -I~/src/ruby/trunk/.ext/x86_64-linux" --repeat-count 1 --repeat-result average -vv benchmark.yml
cold: ruby 2.6.0dev (2018-08-19 coldfunc-init 64464) [x86_64-linux]
last_commit=Let all Init_xxx and extension APIs frequently called from init code paths be considered cold
trunk: ruby 2.6.0dev (2018-08-18 trunk 64448) [x86_64-linux]
Calculating -------------------------------------
                           cold       trunk 
           optcarrot     43.519      43.835 fps

Comparison:
                        optcarrot
               trunk:        43.8 fps 
                cold:        43.5 fps - 1.01x  slower

lourens@CarbonX1:~/src/optcarrot$ benchmark-driver -e "cold::~/src/ruby/ruby/ruby -I~/src/ruby/ruby/lib -I~/src/ruby/ruby/. -I~/src/ruby/ruby/.ext/x86_64-linux" -e "trunk::~/src/ruby/trunk/ruby -I~/src/ruby/trunk/lib -I~/src/ruby/trunk/. -I~/src/ruby/trunk/.ext/x86_64-linux" --repeat-count 1 --repeat-result average -vv benchmark.yml
cold: ruby 2.6.0dev (2018-08-19 coldfunc-init 64464) [x86_64-linux]
last_commit=Let all Init_xxx and extension APIs frequently called from init code paths be considered cold
trunk: ruby 2.6.0dev (2018-08-18 trunk 64448) [x86_64-linux]
Calculating -------------------------------------
                           cold       trunk 
           optcarrot     45.612      41.826 fps

Comparison:
                        optcarrot
                cold:        45.6 fps 
               trunk:        41.8 fps - 1.09x  slower

lourens@CarbonX1:~/src/optcarrot$ benchmark-driver -e "cold::~/src/ruby/ruby/ruby -I~/src/ruby/ruby/lib -I~/src/ruby/ruby/. -I~/src/ruby/ruby/.ext/x86_64-linux" -e "trunk::~/src/ruby/trunk/ruby -I~/src/ruby/trunk/lib -I~/src/ruby/trunk/. -I~/src/ruby/trunk/.ext/x86_64-linux" --repeat-count 1 --repeat-result average -vv benchmark.yml
cold: ruby 2.6.0dev (2018-08-19 coldfunc-init 64464) [x86_64-linux]
last_commit=Let all Init_xxx and extension APIs frequently called from init code paths be considered cold
trunk: ruby 2.6.0dev (2018-08-18 trunk 64448) [x86_64-linux]
Calculating -------------------------------------
                           cold       trunk 
           optcarrot     44.258      43.458 fps

Comparison:
                        optcarrot
                cold:        44.3 fps 
               trunk:        43.5 fps - 1.02x  slower
```